feat: migrate ad banner and announcement banner to fe content
