--- packages/web/config/ads-banner.json
@@ -1,60 +1,0 @@
-{
-  "banners": [
-    {
-      "name": "milkTIA",
-      "start_date": "2023-12-18T14:30:00.123Z",
-      "end_date": "2024-01-25T14:30:00.123Z",
-      "header": "Got Milk?",
-      "subheader": "$milkTIA now available on Osmosis",
-      "external_url": "/pool/1335",
-      "icon_image_url": "https://raw.githubusercontent.com/osmosis-labs/fe-content/main/images/milkTIA.svg",
-      "icon_image_alt": "milkTIA logo",
-      "gradient": "linear-gradient(93.62deg, #F7D1F2 -6.6%, #4B99FC 46.78%, #D3AAFB 105.39%)",
-      "font_color": "#213469",
-      "arrow_color": "#213469",
-      "featured": true
-    },
-    {
-      "name": "Place limit orders",
-      "start_date": "2023-06-22T14:30:00.123Z",
-      "end_date": "2024-06-22T14:30:00.123Z",
-      "header": "Place limit orders",
-      "subheader": "Experience pro trading with TFM",
-      "external_url": "https://osmosis.tfm.com/",
-      "icon_image_url": "https://raw.githubusercontent.com/osmosis-labs/fe-content/main/images/place-limit-order.svg",
-      "icon_image_alt": "Variable chart decreasing to the lower right, then increasing to the upper right with dollar signs",
-      "gradient": "linear-gradient(136deg, #3B154F 0%, #10061C 47.66%, #0E2654 100%)",
-      "font_color": "#E4E1FB",
-      "arrow_color": "#BAF3FF",
-      "featured": true
-    },
-    {
-      "name": "Looking to DCA?",
-      "start_date": "2023-06-22T14:30:00.123Z",
-      "end_date": "2024-06-22T14:30:00.123Z",
-      "header": "Looking to DCA?",
-      "subheader": "Set up recurring swaps with CalcFi",
-      "external_url": "https://app.calculated.fi/create-strategy/dca-in/assets/?chain=Osmosis&utm_source=osmosis_dex",
-      "icon_image_url": "https://raw.githubusercontent.com/osmosis-labs/fe-content/main/images/dollar-cost-averaging.svg",
-      "icon_image_alt": "Chart increasing to the upper right with dollar signs",
-      "gradient": "linear-gradient(136deg, #543E17 0%, #161A24 47.66%, #183B5B 100%)",
-      "font_color": "#FFF5C7",
-      "arrow_color": "#EED75F",
-      "featured": true
-    },
-    {
-      "name": "Margined Protocol",
-      "start_date": "2023-11-10T14:30:00.123Z",
-      "end_date": "2024-11-10T14:30:00.123Z",
-      "header": "Feeling bullish?",
-      "subheader": "Try quadratic leverage on Margined",
-      "external_url": "https://trade.margined.io/",
-      "icon_image_url": "https://raw.githubusercontent.com/osmosis-labs/fe-content/main/images/feeling-bullish.svg",
-      "icon_image_alt": "Chart increasing exponentially to the upper right over a background of dots",
-      "gradient": "linear-gradient(136deg, #020200 0%, #000000 47.66%, #373E24 100%)",
-      "font_color": "#CBFD50",
-      "arrow_color": "#CBFD50",
-      "featured": true
-    }
-  ]
-}
