+++ packages/web/pages/index.tsx
@@ -1,35 +1,17 @@
+import { useQuery } from "@tanstack/react-query";
 import { observer } from "mobx-react-lite";
-import type { GetStaticProps, InferGetStaticPropsType } from "next";
 
-import { Ad, AdCMS } from "~/components/ad-banner/ad-banner-types";
+import { AdBanner } from "~/components/ad-banner";
+import ErrorBoundary from "~/components/error/error-boundary";
 import { ProgressiveSvgImage } from "~/components/progressive-svg-image";
 import { SwapTool } from "~/components/swap-tool";
 import { EventName } from "~/config";
-import adCMSData from "~/config/ads-banner.json";
-import { useAmplitudeAnalytics } from "~/hooks";
+import { useAmplitudeAnalytics, useFeatureFlags } from "~/hooks";
+import { queryGithubFile } from "~/server/queries/github";
 
-interface HomeProps {
-  ads: Ad[];
-}
+const Home = () => {
+  const featureFlags = useFeatureFlags();
 
-export const getStaticProps: GetStaticProps<HomeProps> = async () => {
-  let ads: Ad[] = [];
-
-  const adCMS = adCMSData as AdCMS;
-
-  try {
-    // const { data: adCMS }: { data: AdCMS } = await axiosInstance.get(
-    //   ADS_BANNER_URL
-    // );
-    ads = adCMS.banners.filter(({ featured }) => featured);
-  } catch (error) {
-    console.error("Error fetching ads:", error);
-  }
-
-  return { props: { ads } };
-};
-
-const Home = ({ ads }: InferGetStaticPropsType<typeof getStaticProps>) => {
   useAmplitudeAnalytics({
     onLoadEvent: [EventName.Swap.pageViewed, { isOnHome: true }],
   });
@@ -66,10 +48,50 @@ const Home = ({ ads }: InferGetStaticPropsType<typeof getStaticProps>) => {
       </div>
       <div className="my-auto flex h-auto w-full items-center">
         <div className="ml-auto mr-[15%] flex w-[27rem] flex-col gap-4 lg:mx-auto md:mt-mobile-header">
-          <SwapTool ads={ads} />
+          {featureFlags.swapsAdBanner && <SwapAdsBanner />}
+          <SwapTool />
         </div>
       </div>
     </main>
+  );
+};
+
+export interface Ad {
+  name: string;
+  start_date: string;
+  end_date: string;
+  header: string;
+  subheader: string;
+  external_url: string;
+  icon_image_url: string;
+  icon_image_alt: string;
+  gradient: string;
+  font_color: string;
+  arrow_color: string;
+  featured: true;
+}
+
+const SwapAdsBanner = () => {
+  /**
+   * Fetches the latest update from the osmosis-labs/fe-content repo
+   * @see https://github.com/osmosis-labs/fe-content/blob/main/cms/swap-rotating-banner.json
+   */
+  const { data, isLoading } = useQuery({
+    queryKey: ["swap-ads-banner"],
+    queryFn: () =>
+      queryGithubFile<{ banners: Ad[] }>({
+        repo: "osmosis-labs/fe-content",
+        filePath: "cms/swap-rotating-banner.json",
+      }),
+  });
+
+  if (!data?.banners || isLoading) return null;
+
+  return (
+    // If there is an error, we don't want to show the banner
+    <ErrorBoundary fallback={null}>
+      <AdBanner ads={data.banners} />
+    </ErrorBoundary>
   );
 };
